+++ CHANGELOG.md
@@ -10,6 +10,13 @@ Since there is no stable release yet, the changes are organized per day in
 reverse chronological order. The main purpose of this document in its current
 state is to list breaking changes.
 
+## [2023-04-05]
+
+### Breaking changes
+
+- The `nih_debug_assert*!()` macros are now upgraded to regular panicking
+  `debug_assert!()` macros during tests.
+
 ## [2023-04-01]
 
 ### Fixed